Grammatical fixes.
NOTE! I stuck a comment in the documentation about `# >= 10 AND # <= 20`, because I believe the documentation is incorrect.

Also, if you want to encourage users to create issues for feature requests then a better option for the very last paragraph would be:

If there's some functionality you'd like to see added to JsQuery, please [create a github issue](https://github.com/akorotkov/jsquery/issues). We welcome pull requests, but it would be best to open an issue for discussion first.