Fix build due to new checks in PostgreSQL 16
The macro PG_DETOAST_DATUM returns a pointer to a varlena structure and the
input to the function DatumGetPointer must be of type Datum (see the commit
c8b2ef05f481ef06326d7b9f3eb14b303f215c7e in PostgreSQL 16). So use a simple cast
instead of the function DatumGetPointer.